requirements.txt
certifi
chardet
idna
requests
selenium==4.4.3
urllib3
ipapi
func_timeout
random-word==1.0.7
pyyaml
notifiers
pyvirtualdisplay
webdriver-manager>=3.8.5
python-dotenv
pymongo
pycountry
path